refactor: Group `Benchmark{,Group,Manager}.js` under 📁`Benchmark`
