+++ src/Benchmark/BenchmarkManager.js
